requirements.txt
Flask==1.1.1
requests==2.24.0
waitress==1.4.4
python-dateutil==2.8.1
google-cloud-storage==1.30.0
